+++ hcl2template/types.packer_config_test.go
@@ -110,23 +110,34 @@ func TestParser_complete(t *testing.T) {
 					Builder:  basicMockBuilder,
 					Provisioners: []packer.CoreBuildProvisioner{
 						{
-							PType:       "shell",
-							PName:       "provisioner that does something",
-							Provisioner: basicMockProvisioner,
+							PType: "shell",
+							PName: "provisioner that does something",
+							Provisioner: &HCL2Provisioner{
+								Provisioner: basicMockProvisioner,
+							},
 						},
-						{PType: "file", Provisioner: basicMockProvisioner},
+						{
+							PType: "file",
+							Provisioner: &HCL2Provisioner{
+								Provisioner: basicMockProvisioner,
+							},
+						},
 					},
 					PostProcessors: [][]packer.CoreBuildPostProcessor{
 						{
 							{
-								PType:             "amazon-import",
-								PName:             "something",
-								PostProcessor:     basicMockPostProcessor,
+								PType: "amazon-import",
+								PName: "something",
+								PostProcessor: &HCL2PostProcessor{
+									PostProcessor: basicMockPostProcessor,
+								},
 								KeepInputArtifact: pTrue,
 							},
 							{
-								PType:         "amazon-import",
-								PostProcessor: basicMockPostProcessor,
+								PType: "amazon-import",
+								PostProcessor: &HCL2PostProcessor{
+									PostProcessor: basicMockPostProcessor,
+								},
 							},
 						},
 					},
@@ -146,23 +157,34 @@ func TestParser_complete(t *testing.T) {
 					},
 					Provisioners: []packer.CoreBuildProvisioner{
 						{
-							PType:       "shell",
-							PName:       "provisioner that does something",
-							Provisioner: basicMockProvisioner,
+							PType: "shell",
+							PName: "provisioner that does something",
+							Provisioner: &HCL2Provisioner{
+								Provisioner: basicMockProvisioner,
+							},
 						},
-						{PType: "file", Provisioner: basicMockProvisioner},
+						{
+							PType: "file",
+							Provisioner: &HCL2Provisioner{
+								Provisioner: basicMockProvisioner,
+							},
+						},
 					},
 					PostProcessors: [][]packer.CoreBuildPostProcessor{
 						{
 							{
-								PType:             "amazon-import",
-								PName:             "something",
-								PostProcessor:     basicMockPostProcessor,
+								PType: "amazon-import",
+								PName: "something",
+								PostProcessor: &HCL2PostProcessor{
+									PostProcessor: basicMockPostProcessor,
+								},
 								KeepInputArtifact: pTrue,
 							},
 							{
-								PType:         "amazon-import",
-								PostProcessor: basicMockPostProcessor,
+								PType: "amazon-import",
+								PostProcessor: &HCL2PostProcessor{
+									PostProcessor: basicMockPostProcessor,
+								},
 							},
 						},
 					},