+++ CHANGELOG.md
@@ -3,14 +3,20 @@
 ### IMPROVEMENTS:
 * builder/file: Create parent directories of target file, if they don't exist.
     [GH-9452]
+* builder/google: Add support for oslogin via the `use_os_login` configuration
+    option [GH-9339]
+* builder/googlecompute: Add `wrap_startup_script` configuration option to
+    allow the disabling of Packer's startup script wrapper [GH-9505]
 * builder/hyperone: Add support for custom username in vm creation. [GH-9497]
 * builder/hyperone: Skip chroot device discovery. [GH-9489]
 * builder/proxmox: Enable Proxmox builder to toggle the firewall parameter for
     network interfaces. [GH-9487]
 * builder/proxmox: Update Proxmox storagePoolTypes [GH-9418]
+* builder/qemu: Add 'cdrom_interface' option to QEMU builder [GH-9483]
 * builder/ucloud: New access config options and run config options. [GH-9466]
 * builder/vsphere-clone: Add `boot_command` support to vsphere-clone builder,
     including support for starting an HTTP server
+* builder/vsphere-clone: Add `vapp` configuration option [GH-9507]
 * builder/vsphere: Add boot_command support to vsphere-clone builder. [GH-9397]
 * builder/vsphere: Create vm output folders if they don't exist [GH-9402]
 * builder/vsphere: Look at all available IPs in the waiting for IP loop.
@@ -31,7 +37,7 @@
     types. [GH-9454]
 * core/hcl: Allow use of `keep_input_artifact` in post processors. [GH-9477]
 * core/hcl: Share build info with Provisioner and Post-Processor via HCL2
-    variables [GH-9444]
+    variables [GH-9444] [GH-9534]
 * core: Add on-error flag option to run error-cleanup-provisioner [GH-9429]
 * post-processor/yandex-export: Support Authentication by Service Account Key
     file [GH-9379]
@@ -55,12 +61,16 @@
     [GH-9387]
 * builder/vmware: update vendor library, enabling retries on 502 errors
     [GH-9391]
+* builder/vsphere-clone: Fix SSH public key injection for cloned templates
+    [GH-9507]
 * builder/vsphere: Fix vsphere ToolsSyncTime and ToolsUpgradePolicy [GH-9515]
 * communicator/winrm: add the "no_proxy" environment variable option to have
     winrm bypass the proxy set by the http_proxy or https_proxy environment
     vars, when connecting to the remote instance. [GH-9267]
 * core: Do not print download progress bar if a machine-readable UI is chosen.
     [GH-9448]
+* post-processor/amazon-import: Add support for retrying RequestLimitExceeded
+    errors when importing an image [GH-9537]
 * provisioner/ansible: Correct check for whether PackerHttpAddr is implemented
     or not [GH-9498]
 
@@ -261,6 +271,11 @@
 
 ## 1.5.5 (March 25,2020)
 
+### BACKWARDS INCOMPATIBILITIES:
+* core: Interpolation of undefined variables will now error as expected, in
+    previous versions variables were allowed to be set as a command line
+    argument but that was because the validation was being ignored
+
 ### IMPROVEMENTS:
 * builder/azure: Add support for configurable KeyVault SKU [GH-8879]
 * builder/hyperv: Add `first_boot_device` setting to allow the selection of the
@@ -295,8 +310,6 @@
     strings [GH-8829]
 
 ### Bug Fixes:
-* bilder/proxmox: Bump proxmox-api-go to fix upstream bug where users hit open
-    file limit. [GH-8800]
 * builder/azure: Fix `winrm_password` attribution and allow users to set
     `winrm_username` [GH-8928]
 * builder/azure: Fix azure key vault cleanup failure [GH-8905]
@@ -307,6 +320,8 @@
     to fix SSH authentication issue [GH-8942]
 * builder/proxmox: Add new validation to catch that template_name cannot
     contain spaces. [GH-8799]
+* builder/proxmox: Bump proxmox-api-go to fix upstream bug where users hit open
+    file limit. [GH-8800]
 * builder/vagrant: Fix path validation in ssh config step. [GH-8826]
 * builder/virtualbox-vm: Fix crash when VM has no snapshots. [GH-8906]
 * builder/virtualbox: Remove all floppy controllers before adding a new one.
@@ -323,10 +338,12 @@
     provisioners. [GH-8771]
 * core: Fix bug where user var recursion could fail intermittently when used
     with env vars [GH-8875]
+* core: Interpolation of undefined variables will now error as expected
 * plugins: Make plugin discovery stricter with respect to periods so that users
     can disable plugins by renaming the extension [GH-8735]
+* provisioner/salt: Fix `no_exit_on_failure` config to work correctly as
+    expected. [GH-9119]
 * provisioner/shell: "inline" config option is now a template engine. [GH-8883]
-* provisioner/salt: Fix `no_exit_on_failure` config to work correctly as expected. [GH-9119]
 
 ## 1.5.4 (February 14, 2020)
 no-change release to fix code-signing on OSX binaries. Since checksums for these